Add alternative English names for two countries.
In the case of the Republic of The Gambia, that's its official name, but Mapbox use "The Gambia" so it's useful to have both. "Tanzania" is a common synonym for the United Republic of Tanzania; again Mapbox uses this in its geocoding API. My motivation is that I'm trying to transform country names returned from the Mapbox API into ISO codes, hence adding the names used over there.